Merged PR 2911: Merge rpt/1712-enable-emailing-scheduled-vendor-extracts to develop
By returning false regardless, Exago will continue execution (by emailing or not emailing when selected.)  Also, the MoveFileToRepo() => CopyFileToRepo().  It appears that someone was trying to delete the file in the the working directory.  Exago cannot find said file if it is manually deleted.  Also, the scheduler deletes the file after execution anyways.

Related work items: #1712